DTSCCI-1461 - Refactoring to fix functional tests
