+++ src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/AddDefendantLitigationFriendNotificationHandler.java
@@ -5,10 +5,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.civil.handler.callback.camunda.notification.NotificationData;
 import uk.gov.hmcts.reform.civil.model.CaseData;
-import uk.gov.hmcts.reform.civil.notify.NotificationService;
-import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
-import uk.gov.hmcts.reform.civil.service.OrganisationService;
-import uk.gov.hmcts.reform.civil.service.flowstate.IStateFlowEngine;
 
 import java.util.HashMap;
 import java.util.HashSet;
@@ -24,22 +20,6 @@ import static uk.gov.hmcts.reform.civil.utils.NotificationUtils.getLegalOrganiza
 @Slf4j
 @RequiredArgsConstructor
 public class AddDefendantLitigationFriendNotificationHandler extends NotificationHandler implements NotificationData {
-
-    protected final NotificationService notificationService;
-    protected final NotificationsProperties notificationsProperties;
-    protected final OrganisationService organisationService;
-    protected final IStateFlowEngine stateFlowEngine;
-
-    protected static final String REFERENCE_TEMPLATE_APPLICANT = "litigation-friend-added-applicant-notification-%s";
-    protected static final String REFERENCE_TEMPLATE_RESPONDENT = "litigation-friend-added-respondent-notification-%s";
-
-    @Override
-    protected void sendNotification(Set<EmailDTO> recipients) {
-        for (EmailDTO recipient : recipients) {
-            notificationService.sendMail(recipient.getTargetEmail(), recipient.getEmailTemplate(), recipient.getParameters(),
-                    recipient.getReference());
-        }
-    }
 
     @Override
     public Map<String, String> addProperties(CaseData caseData) {
